# apps/wanaku-router-backend/src/main/webui/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "ForwardReference" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "namespace" : {
            "type" : "string"
          }
        }
      },
      "ForwardWithRootsRequest" : {
        "type" : "object",
        "properties" : {
          "forward" : {
            "$ref" : "#/components/schemas/ForwardReference"
          },
          "roots" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RootEntry"
            }
          }
        }
      },
...
      "Param" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        }
      },
...
      "ResourceReference" : {
        "type" : "object",
        "properties" : {
          "labels" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "id" : {
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "mimeType" : {
            "type" : "string"
          },
          "params" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Param"
            }
          },
          "configurationURI" : {
            "type" : "string"
          },
          "secretsURI" : {
            "type" : "string"
          },
          "namespace" : {
            "type" : "string"
          }
        }
      },
      "RootEntry" : {
        "type" : "object",
        "properties" : {
          "uri" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
      "WanakuError" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "WanakuResponseForwardReference" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "$ref" : "#/components/schemas/ForwardReference"
          }
        }
      },
...
      "WanakuResponseListForwardReference" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ForwardReference"
            }
          }
        }
      },
...
      "WanakuResponseResourceReference" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "$ref" : "#/components/schemas/ResourceReference"
          }
        }
      },
...
      "WanakuResponseVoid" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : { }
        }
      }
    }
  },
  "paths" : {
...
    "/api/v1/forwards" : {
      "get" : {
        "parameters" : [ {
          "name" : "labelFilter",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseListForwardReference"
                }
              }
            }
          }
        },
        "summary" : "List Forwards",
        "tags" : [ "Forwards Resource" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ForwardReference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Add Forward",
        "tags" : [ "Forwards Resource" ]
      }
    },
    "/api/v1/forwards/with-roots" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ForwardWithRootsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Add Forward With Roots",
        "tags" : [ "Forwards Resource" ]
      }
    },
    "/api/v1/forwards/{name}" : {
      "put" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ForwardReference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update",
        "tags" : [ "Forwards Resource" ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          }
        },
        "summary" : "Remove Forward",
        "tags" : [ "Forwards Resource" ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseForwardReference"
                }
              }
            }
          }
        },
        "summary" : "Get By Name",
        "tags" : [ "Forwards Resource" ]
      }
    },
    "/api/v1/forwards/{name}/refreshes" : {
      "post" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          }
        },
        "summary" : "Refresh",
        "tags" : [ "Forwards Resource" ]
      }
    },
...
    "/api/v1/resources/{name}" : {
      "put" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResourceReference"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update",
        "tags" : [ "Resources Resource" ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          }
        },
        "summary" : "Remove",
        "tags" : [ "Resources Resource" ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseResourceReference"
                }
              }
            }
          }
        },
        "summary" : "Get By Name",
        "tags" : [ "Resources Resource" ]
      }
    },
...
  },
  "info" : {
    "title" : "wanaku-router-backend API",
    "version" : "0.3.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}