# core/core-services-api/src/main/java/ai/wanaku/core/services/api/ForwardsService.java
package ai.wanaku.core.services.api;

import jakarta.ws.rs.Consumes;
import jakarta.ws.rs.DELETE;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.PUT;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.MediaType;

import java.util.List;
import java.util.stream.Collectors;
import ai.wanaku.capabilities.sdk.api.types.ForwardReference;
import ai.wanaku.capabilities.sdk.api.types.WanakuResponse;

/**
 * JAX-RS service interface for managing forward references in the Wanaku system.
 * <p>
 * This service provides REST endpoints for managing forwards, which enable the
 * Wanaku router to proxy requests to remote MCP (Model Context Protocol) servers
 * or other external capability providers.
 * <p>
 * All endpoints are available under the {@code /api/v1/forwards} base path.
 */
@Path("/api/v1/forwards")
public interface ForwardsService {
    /**
     * Registers a new forward reference in the system.
     * <p>
     * A forward reference configures the router to proxy requests for specific
     * capabilities to a remote server or service.
     *
     * @param reference the forward reference to register
     * @return a {@link WanakuResponse} indicating the result of the add operation
     */
    @POST
    @Produces(MediaType.APPLICATION_JSON)
    @Consumes(MediaType.APPLICATION_JSON)
    WanakuResponse<Void> addForward(ForwardReference reference);

    /**
     * Removes a forward reference from the system.
     *
     * @param name the name of the forward to remove
     * @return a {@link WanakuResponse} indicating the result of the remove operation
     */
    @Path("/{name}")
    @DELETE
    @Produces(MediaType.APPLICATION_JSON)
    WanakuResponse<Void> removeForward(@PathParam("name") String name);

    /**
     * Lists all registered forward references.
     *
     * @param labelFilter optional label expression to filter forwards by labels
     * @return a {@link WanakuResponse} containing a list of all forward references
     */
    @GET
    @Produces(MediaType.APPLICATION_JSON)
    WanakuResponse<List<ForwardReference>> listForwards(@QueryParam("labelFilter") String labelFilter);

    /**
     * Lists all registered forward references without filtering.
     *
     * @return a {@link WanakuResponse} containing a list of all forward references
     */
    default WanakuResponse<List<ForwardReference>> listForwards() {
        return listForwards(null);
    }

    /**
     * Retrieves a forward reference by its name.
     *
     * @param name the name of the forward to retrieve
     * @return a {@link WanakuResponse} containing the forward reference
     */
    @Path("/{name}")
    @GET
    @Produces(MediaType.APPLICATION_JSON)
    WanakuResponse<ForwardReference> getByName(@PathParam("name") String name);

    /**
     * Updates an existing forward reference.
     *
     * @param name the name of the forward to update
     * @param reference the updated forward reference
     * @return a {@link WanakuResponse} indicating the result of the update operation
     */
    @Path("/{name}")
    @PUT
    @Consumes(MediaType.APPLICATION_JSON)
    @Produces(MediaType.APPLICATION_JSON)
    WanakuResponse<Void> updateForward(@PathParam("name") String name, ForwardReference reference);

    /**
     * Refreshes a forward reference to re-discover tools and resources from the remote server.
     *
     * @param name the name of the forward to refresh
     * @return a {@link WanakuResponse} indicating the result of the refresh operation
     */
    @Path("/{name}/refreshes")
    @POST
    @Produces(MediaType.APPLICATION_JSON)
    WanakuResponse<Void> refreshForward(@PathParam("name") String name);

    /**
     * Registers a new forward reference with MCP roots configuration.
     * <p>
     * The roots are directories or resources that the upstream MCP server is allowed to access.
     * Some servers (e.g., the filesystem server) require at least one root and will refuse to
     * operate if the client returns an empty roots list.
     *
     * @param request the forward reference with roots configuration
     * @return a {@link WanakuResponse} indicating the result of the add operation
     */
    @Path("/with-roots")
    @POST
    @Produces(MediaType.APPLICATION_JSON)
    @Consumes(MediaType.APPLICATION_JSON)
    WanakuResponse<Void> addForwardWithRoots(ForwardWithRootsRequest request);

    /**
     * Convenience method to register a forward with roots specified as URI strings.
     *
     * @param reference the forward reference to register
     * @param rootUris  the list of root URIs to advertise to the upstream server
     * @return a {@link WanakuResponse} indicating the result of the add operation
     */
    default WanakuResponse<Void> addForwardWithRoots(ForwardReference reference, List<String> rootUris) {
        ForwardWithRootsRequest request = new ForwardWithRootsRequest();
        request.setForward(reference);
        if (rootUris != null) {
            request.setRoots(rootUris.stream()
                    .map(uri -> new ForwardWithRootsRequest.RootEntry(uri, null))
                    .collect(Collectors.toList()));
        }
        return addForwardWithRoots(request);
    }
}